Updates for Kubernetes 1.26
Kubernetes 1.26 requires a container runtime implementing
the V1 CRI API; which means containerd 1.6; which is not
available in Ubuntu LTS at the moment.

Fix:
- add Docker upstream APT repo (which has containerd 1.6)
- install Docker Engine and containerd from that repo

We don't really need the Docker Engine but it also doesn't
hurt and it's convenient to be able to do a quick "docker run"
just in case. Also, it might help with autoloading some
modules and enabling IP forwarding (not 100% sure about that,
but in my experiments when just installing "bare" containerd
there are a number of things that are not set properly).