Rework the mechanism to wait for the control plane
Instead of doing a "curl" on the control plane
API server (which is simple, but doesn't tell us
what's going on and is hard to troubleshoot when
something goes wrong), we now use a provisioner
to wait for cloud-init to complete.

Furthermore, the provisioner tails the cloud-init
logs so that we can see in real time what's going on.